Added mouse controls, moved div by 100 logic into controller and out of widget

+++ GLPolygon.h
@@ -5,9 +5,9 @@ class GLPolygon
 {
     public:       
         int vertices;
-        int x;
-        int y;
-        int angle;
+        float x;
+        float y;
+        float angle;
         
         GLPolygon(int vertices = 3);
         void render();
